# api/swagger/swagger.yaml
# this is an example of the Uber API
# as a demonstration of an API spec in YAML
swagger: '2.0'
info:
  title: Copy-and-Paste
  version: "0.0.1"
# array of all schemes that your API supports
schemes:
  - https
  - http
basePath: /api
produces:
  - application/json
paths:
  /connect/{id}:
    x-swagger-router-controller: connect
    post:
      operationId: connectById
      summary: Connects to client specified by ID
      description: |
        This function will reserve the client with specified ID to you. No other client can communicate with this client anymore
      tags:
        - connect
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: Succesfully connected
          schema:
            $ref: '#/definitions/Success'
        404:
          description: ID is not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error occured
          schema:
            $ref: '#/definitions/Error'
  /send/{id}:
    x-swagger-router-controller: send
    post:
      operationId: sendById
      summary: Sends information to web-client
      description: With this method you can send data directly to web-client. The client is specified with same ID that in connect
      tags:
        - send
      parameters:
        - name: id
          in: path
          description: The ID of the web-client
          required: true
          type: string
        - name: body
          in: body
          description: Message body
          schema:
            $ref: '#/definitions/PostData'
          required: true
      responses:
        200:
          description: The data has been succesfully send and received by the web-client
          schema:
            $ref: '#/definitions/Success'
        400:
          description: The message in body is missing
          schema:
            $ref: '#/definitions/Error'
        404:
          description: ID is not found
          schema:
            $ref: '#/definitions/Error'
        408:
          description: Web-client did not response to this message in specified time. You may try to send message again
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error occured
          schema:
            $ref: '#/definitions/Error'
definitions:
  PostData:
    type: object
    required:
      - data
    properties:
      data: 
        type: string
  Success:
    type: object
    properties:
      msg:
        type: string
  Error:
    type: object
    properties:
      msg:
        type: string